# Replace factory functions with classes.