@udifuchs 2025-02-10
> @udifuchs any update on this?

I just updated my branch to be in sync with lxml:master. But I don't get to decide if it is going to be merged.

When I made this PR about a year ago, I had plans to publish a matching PR to types-lxml once this PR is accepted. Meanwhile, types-lxml went through a lot of changes. Updating my changes there is a bit more than just fixing conflicts. I will try to update it in the next few days and send it to types-lxml.

Generally, these two PR should be independent. But I don't think that it  make sense make Element a class in types-lxml, unless lxml makes this change first.

Also, types-lxml does some runtime checks of the types, but I still don't fully understand how these work and if it has any implications for my change.