@udifuchs 2025-02-13
I created a PR for the types-lxml: https://github.com/abelcheung/types-lxml/pull/81

If the pull request to types-lxml is applied, users will be able to annotate using the public Element "class":
```
def print_tree(tree: e.ElementTree[e.Element]) -> None: ...
```
It would work even if this PR is not applied.

But without this PR, users will still need to use the private _Element for  type narrowing:
```
def print_element(element: e.Element) -> None: ...

def func(element: object) -> None:
    # isinstance(element, e.Element)   # This line will raise an Exception without this PR.
    isinstance(element, e._Element)  # The type of element is now narrowed to e._Element
    print_element(element)  # Function call is valid static typing
```