@dpgraham4401 2025-02-13
Thanks @udifuchs, I wasn't aware of the [types-lxml](https://github.com/abelcheung/types-lxml) package, that's very helpful.