@scoder 2025-02-17
I ran the [etree benchmark](https://github.com/lxml/lxml/blob/f56aa1de81a644189d6c4bbb3495043cdacc0500/benchmark/bench_etree.py) over your branch on Python 3.12. The tree setup times grow by 5-25%:
```
Current master:
     T1: 0.0155 0.0170 0.0177 0.0152 0.0172 0.0174
     T2: 0.0164 0.0191 0.0193 0.0184 0.0208 0.0215
     T3: 0.0036 0.0045 0.0040 0.0146 0.0154 0.0154
     T4: 0.0001 0.0001 0.0001 0.0003 0.0003 0.0003

lxe: create_elements              (--TC T1   )    0.0286 msec/pass, best of (   0.0288    0.0286    0.0291)
lxe: create_elements              (--TC T2   )    0.5600 msec/pass, best of (   0.5641    0.5600    0.5634)
lxe: create_elements              (--TC T3   )    0.0031 msec/pass, best of (   0.0031    0.0033    0.0036)
lxe: create_elements              (--TC T4   )    0.0281 msec/pass, best of (   0.0281    0.0281    0.0281)
lxe: create_subelements           (--TC T1   )    0.0253 msec/pass, best of (   0.0260    0.0253    0.0255)
lxe: create_subelements           (--TC T2   )    0.4935 msec/pass, best of (   0.4983    0.4935    0.4942)
lxe: create_subelements           (--TC T3   )    0.0029 msec/pass, best of (   0.0029    0.0031    0.0031)
lxe: create_subelements           (--TC T4   )    0.0253 msec/pass, best of (   0.0257    0.0253    0.0260)

PR branch:
     T1: 0.0161 0.0183 0.0182 0.0158 0.0188 0.0186
     T2: 0.0174 0.0232 0.0217 0.0234 0.0253 0.0247
     T3: 0.0040 0.0047 0.0045 0.0172 0.0170 0.0165
     T4: 0.0001 0.0001 0.0001 0.0003 0.0003 0.0004

lxe: create_elements              (--TC T1   )    0.0298 msec/pass, best of (   0.0298    0.0298    0.0300)
lxe: create_elements              (--TC T2   )    0.5920 msec/pass, best of (   0.5949    0.5941    0.5920)
lxe: create_elements              (--TC T3   )    0.0036 msec/pass, best of (   0.0036    0.0036    0.0036)
lxe: create_elements              (--TC T4   )    0.0293 msec/pass, best of (   0.0296    0.0298    0.0293)
lxe: create_subelements           (--TC T1   )    0.0272 msec/pass, best of (   0.0272    0.0279    0.0272)
lxe: create_subelements           (--TC T2   )    0.5312 msec/pass, best of (   0.5312    0.5329    0.5319)
lxe: create_subelements           (--TC T3   )    0.0031 msec/pass, best of (   0.0031    0.0033    0.0033)
lxe: create_subelements           (--TC T4   )    0.0269 msec/pass, best of (   0.0269    0.0269    0.0269)
```

Given that the goal of this PR is pretty much code cosmetics, I find a performance loss in that order rather high for a library that aims for high-speed XML processing.