@scoder 2025-02-18
I was surprised myself that the numbers were that bad, given that we're only changing `Element()` and not `SubElement()`, which should largely dominate the tree building. So I triple checked and found that I had run the benchmarks in debug mode ~ sorry for that.

I did the timings again and it looks much better:
```
Current master:
     T1: 0.0160 0.0153 0.0149 0.0132 0.0149 0.0161
     T2: 0.0150 0.0164 0.0172 0.0169 0.0192 0.0192
     T3: 0.0034 0.0036 0.0034 0.0148 0.0150 0.0160
     T4: 0.0001 0.0001 0.0001 0.0005 0.0004 0.0003

lxe: create_elements              (--TC T1   )    0.0215 msec/pass, best of (   0.0224    0.0224    0.0215)
lxe: create_elements              (--TC T2   )    0.4103 msec/pass, best of (   0.4108    0.4103    0.4120)
lxe: create_elements              (--TC T3   )    0.0021 msec/pass, best of (   0.0024    0.0024    0.0021)
lxe: create_elements              (--TC T4   )    0.0205 msec/pass, best of (   0.0212    0.0207    0.0205)
lxe: create_subelements           (--TC T1   )    0.0215 msec/pass, best of (   0.0215    0.0310    0.0231)
lxe: create_subelements           (--TC T2   )    0.4170 msec/pass, best of (   0.4199    0.4170    0.4344)
lxe: create_subelements           (--TC T3   )    0.0024 msec/pass, best of (   0.0024    0.0029    0.0024)
lxe: create_subelements           (--TC T4   )    0.0215 msec/pass, best of (   0.0215    0.0215    0.0215)

PR branch:
     T1: 0.0137 0.0155 0.0159 0.0136 0.0155 0.0162
     T2: 0.0156 0.0174 0.0180 0.0180 0.0195 0.0204
     T3: 0.0034 0.0038 0.0037 0.0147 0.0149 0.0151
     T4: 0.0001 0.0001 0.0001 0.0003 0.0003 0.0003

lxe: create_elements              (--TC T1   )    0.0231 msec/pass, best of (   0.0236    0.0231    0.0234)
lxe: create_elements              (--TC T2   )    0.4590 msec/pass, best of (   0.4933    0.4590    0.4714)
lxe: create_elements              (--TC T3   )    0.0052 msec/pass, best of (   0.0052    0.0052    0.0055)
lxe: create_elements              (--TC T4   )    0.0236 msec/pass, best of (   0.0455    0.0238    0.0236)
lxe: create_subelements           (--TC T1   )    0.0229 msec/pass, best of (   0.0229    0.0231    0.0234)
lxe: create_subelements           (--TC T2   )    0.4358 msec/pass, best of (   0.4554    0.4358    0.4492)
lxe: create_subelements           (--TC T3   )    0.0041 msec/pass, best of (   0.0041    0.0052    0.0052)
lxe: create_subelements           (--TC T4   )    0.0219 msec/pass, best of (   0.0241    0.0241    0.0219)
```

So, the regression in tree building (the first block of timings) is always in the order of ~5%, but you can see the big difference for `Element()` in the `create_elements` micro-benchmark, which is much slower now (> 20%).

I think a 5% slow-down in realistic code is probably acceptable for this, but the regression is still rather unfortunate. There's certainly code out there that will see a real difference.