@udifuchs 2025-02-18
The difference between master and this branch is that instead of calling the `Element` function, cpython is calling the `Element.__new__` class method. My guess is that the slowdown is from cpython having to inspect the `Element` class.

If cython classes could have a `__new__` method, it might have been able to accelerate this. But I don't have a deep understanding of cython. I guess that you are the expert...