@scoder 2025-02-18
I tried the benchmark again, this time using Cython 3.1a1+ (current master branch). For the previous runs I used the (almost latest) 3.0.11 stable release 

The results look much better again:
```
Current master:
     T1: 0.0151 0.0155 0.0155 0.0137 0.0156 0.0157
     T2: 0.0155 0.0168 0.0171 0.0175 0.0203 0.0215
     T3: 0.0047 0.0039 0.0035 0.0162 0.0149 0.0153
     T4: 0.0001 0.0001 0.0001 0.0003 0.0003 0.0003

lxe: create_elements              (--TC T1   )    0.0217 msec/pass, best of (   0.0217    0.0219    0.0217)
lxe: create_elements              (--TC T2   )    0.4268 msec/pass, best of (   0.4375    0.4268    0.4303)
lxe: create_elements              (--TC T3   )    0.0029 msec/pass, best of (   0.0029    0.0029    0.0029)
lxe: create_elements              (--TC T4   )    0.0217 msec/pass, best of (   0.0417    0.0224    0.0217)
lxe: create_subelements           (--TC T1   )    0.0212 msec/pass, best of (   0.0215    0.0215    0.0212)
lxe: create_subelements           (--TC T2   )    0.4232 msec/pass, best of (   0.4239    0.4590    0.4232)
lxe: create_subelements           (--TC T3   )    0.0024 msec/pass, best of (   0.0024    0.0024    0.0026)
lxe: create_subelements           (--TC T4   )    0.0212 msec/pass, best of (   0.0215    0.0212    0.0212)

PR branch:
     T1: 0.0133 0.0151 0.0155 0.0132 0.0148 0.0151
     T2: 0.0148 0.0170 0.0172 0.0167 0.0191 0.0192
     T3: 0.0030 0.0039 0.0037 0.0143 0.0149 0.0144
     T4: 0.0001 0.0001 0.0001 0.0005 0.0003 0.0003

lxe: create_elements              (--TC T1   )    0.0362 msec/pass, best of (   0.0362    0.0424    0.0424)
lxe: create_elements              (--TC T2   )    0.4439 msec/pass, best of (   0.4439    0.4444    0.4442)
lxe: create_elements              (--TC T3   )    0.0029 msec/pass, best of (   0.0029    0.0029    0.0029)
lxe: create_elements              (--TC T4   )    0.0243 msec/pass, best of (   0.0248    0.0243    0.0248)
lxe: create_subelements           (--TC T1   )    0.0236 msec/pass, best of (   0.0236    0.0236    0.0241)
lxe: create_subelements           (--TC T2   )    0.4199 msec/pass, best of (   0.4265    0.4199    0.4344)
lxe: create_subelements           (--TC T3   )    0.0024 msec/pass, best of (   0.0024    0.0026    0.0024)
lxe: create_subelements           (--TC T4   )    0.0212 msec/pass, best of (   0.0219    0.0212    0.0219)
```
Here, the results are much more mixed and definitely in the same ball park. That's most likely due to the improved support for the Python vectorcall procotol in Cython 3.1, specifically for calls with keyword arguments and kwargs dicts.

I'll postpone the next lxml release (6.0) a little more and try to get out Cython 3.1b1 in the meantime, so that I can base lxml 6.0 at least on that.